buffer is now passed into Raii::get_data
