+++ src/lib.rs
@@ -38,6 +38,7 @@ mod statement;
 use odbc_object::OdbcObject;
 use raii::Raii;
 use result::Return;
+use std::ptr::null_mut;
 pub use diagnostics::{DiagnosticRecord, GetDiagRec};
 pub use result::{Result, EnvAllocError};
 pub use environment::*;
@@ -49,4 +50,21 @@ pub trait Handle {
     type To;
     /// Returns a valid handle to the odbc type.
     unsafe fn handle(&self) -> *mut Self::To;
+}
+
+fn as_out_buffer(buffer: &mut [u8]) -> *mut u8 {
+    if buffer.len() == 0 {
+        null_mut()
+    } else {
+        buffer.as_mut_ptr()
+    }
+}
+
+fn as_buffer_length(n: usize) -> ffi::SQLSMALLINT {
+    use std;
+    if n > std::i16::MAX as usize {
+        std::i16::MAX
+    } else {
+        n as i16
+    }
 }
