+++ src/statement.rs
@@ -1,6 +1,6 @@
 
-use super::{ffi, DataSource, Return, Result, Raii, Handle, Connected};
-use super::ffi::SQLRETURN::*;
+use {ffi, DataSource, Return, Result, Raii, Handle, Connected};
+use ffi::SQLRETURN::*;
 use std::marker::PhantomData;
 use std::ptr::null_mut;
 use std;
@@ -22,6 +22,7 @@ pub struct Statement<'a, S> {
 /// Used to retrieve data from the fields of a query resul
 pub struct Cursor<'a, 'b: 'a> {
     stmt: &'a mut Statement<'b, Executed>,
+    buffer: [u8; 512],
 }
 
 impl<'a, S> Handle for Statement<'a, S> {
@@ -77,7 +78,10 @@ impl<'a> Statement<'a, Executed> {
     /// Returns false on the last row
     pub fn fetch<'b>(&'b mut self) -> Result<Option<Cursor<'b, 'a>>> {
         if self.raii.fetch().into_result(self)? {
-            Ok(Some(Cursor { stmt: self }))
+            Ok(Some(Cursor {
+                        stmt: self,
+                        buffer: [0u8; 512],
+                    }))
         } else {
             Ok(None)
         }
@@ -87,7 +91,7 @@ impl<'a> Statement<'a, Executed> {
 impl<'a, 'b> Cursor<'a, 'b> {
     /// Retrieves data for a single column in the result set
     pub fn get_data(&mut self, col_or_param_num: u16) -> Result<Option<String>> {
-        self.stmt.raii.get_data(col_or_param_num).into_result(self.stmt)
+        self.stmt.raii.get_data(col_or_param_num, &mut self.buffer).into_result(self.stmt)
     }
 }
 
@@ -110,10 +114,10 @@ impl Raii<ffi::Stmt> {
             panic!("Statement text too long");
         }
         match unsafe {
-            ffi::SQLExecDirect(self.handle(),
-                               statement_text.as_ptr(),
-                               length as ffi::SQLINTEGER)
-        } {
+                  ffi::SQLExecDirect(self.handle(),
+                                     statement_text.as_ptr(),
+                                     length as ffi::SQLINTEGER)
+              } {
             ffi::SQL_SUCCESS => Return::Success(true),
             ffi::SQL_SUCCESS_WITH_INFO => Return::SuccessWithInfo(true),
             ffi::SQL_ERROR => Return::Error,
@@ -157,26 +161,31 @@ impl Raii<ffi::Stmt> {
         }
     }
 
-    fn get_data(&mut self, col_or_param_num: u16) -> Return<Option<String>> {
+    fn get_data(&mut self, col_or_param_num: u16, buffer: &mut [u8]) -> Return<Option<String>> {
+        if buffer.len() == 0 {
+            panic!("buffer length may not be null");
+        }
+        if buffer.len() > ffi::SQLLEN::max_value() as usize {
+            panic!("buffer is larger than {} bytes", ffi::SQLLEN::max_value());
+        }
+
         let mut indicator: ffi::SQLLEN = 0;
-        const BUF_LENGTH: ffi::SQLLEN = 512;
-        let mut buf = [0u8; BUF_LENGTH as usize];
         unsafe {
             // Get buffer length...
             let result = ffi::SQLGetData(self.handle(),
                                          col_or_param_num,
                                          ffi::SQL_C_CHAR,
-                                         buf.as_mut_ptr() as ffi::SQLPOINTER,
-                                         BUF_LENGTH,
+                                         buffer.as_mut_ptr() as ffi::SQLPOINTER,
+                                         buffer.len() as ffi::SQLLEN,
                                          &mut indicator as *mut ffi::SQLLEN);
             match result {
                 ffi::SQL_SUCCESS => {
                     if indicator == ffi::SQL_NULL_DATA {
                         Return::Success(None)
                     } else {
-                        Return::Success(Some(std::str::from_utf8(&buf[..(indicator as usize)])
-                            .unwrap()
-                            .to_owned()))
+                        Return::Success(Some(std::str::from_utf8(&buffer[..(indicator as usize)])
+                                                 .unwrap()
+                                                 .to_owned()))
                     }
                 }
                 ffi::SQL_SUCCESS_WITH_INFO => {
@@ -185,24 +194,22 @@ impl Raii<ffi::Stmt> {
                     } else {
                         // Check if string has been truncated. String is also truncated if
                         // indicator is equal to BUF_LENGTH because of terminating nul
-                        if indicator >= BUF_LENGTH {
-                            let extra_space = (indicator as usize + 1) - (BUF_LENGTH as usize - 1);
+                        if indicator >= buffer.len() as ffi::SQLLEN {
+                            let extra_space = (indicator as usize + 1) - (buffer.len() - 1);
                             let mut heap_buf = Vec::with_capacity((indicator as usize) + 1);
                             // Copy everything but the terminating zero into the new buffer
-                            heap_buf.extend_from_slice(&buf[..(BUF_LENGTH as usize) - 1]);
+                            heap_buf.extend_from_slice(&buffer[..(buffer.len() - 1)]);
                             // increase length
                             heap_buf.extend(std::iter::repeat(0).take(extra_space));
                             // Get remainder of string
-                            let ret =
-                                ffi::SQLGetData(self.handle(),
-                                                col_or_param_num,
-                                                ffi::SQL_C_CHAR,
-                                                heap_buf.as_mut_slice()[(BUF_LENGTH as usize) -
-                                                1..]
-                                                    .as_mut_ptr() as
-                                                ffi::SQLPOINTER,
-                                                extra_space as ffi::SQLLEN,
-                                                null_mut());
+                            let ret = ffi::SQLGetData(self.handle(),
+                                                      col_or_param_num,
+                                                      ffi::SQL_C_CHAR,
+                                                      heap_buf.as_mut_slice()[buffer.len() - 1..]
+                                                          .as_mut_ptr() as
+                                                      ffi::SQLPOINTER,
+                                                      extra_space as ffi::SQLLEN,
+                                                      null_mut());
                             heap_buf.pop();
                             let value = String::from_utf8(heap_buf).unwrap();
                             match ret {
@@ -213,10 +220,10 @@ impl Raii<ffi::Stmt> {
                             }
                         } else {
                             // No truncation. Warning may be due to some other issue.
-                            Return::SuccessWithInfo(Some(std::str::from_utf8(&buf[..(indicator as
+                            Return::SuccessWithInfo(Some(std::str::from_utf8(&buffer[..(indicator as
                                                                                  usize)])
-                                .unwrap()
-                                .to_owned()))
+                                                                 .unwrap()
+                                                                 .to_owned()))
                         }
                     }
                 }
@@ -227,3 +234,4 @@ impl Raii<ffi::Stmt> {
         }
     }
 }
+
